Add turbo update partial and test
