+++ test/controllers/ajax_controller_test.rb
@@ -292,36 +292,4 @@ class AjaxControllerTest < FunctionalTestCase
     assert(@response.body.include?(item),
            "Expected #{@response.body} to include #{item}.")
   end
-
-  def test_visual_group_flip_status
-    login
-    visual_group = visual_groups(:visual_group_one)
-    image = images(:agaricus_campestris_image)
-    vgi = visual_group.visual_group_images.find_by(image_id: image.id)
-    new_status = !vgi.included
-    get(:visual_group_status,
-        params: { id: visual_group.id, imgid: image.id, value: new_status })
-    vgi.reload
-    assert_equal(new_status, vgi.included)
-  end
-
-  def test_visual_group_delete_relationship
-    login
-    visual_group = visual_groups(:visual_group_one)
-    image = images(:agaricus_campestris_image)
-    count = VisualGroupImage.count
-    get(:visual_group_status,
-        params: { id: visual_group.id, imgid: image.id, value: "" })
-    assert_equal(count - 1, VisualGroupImage.count)
-  end
-
-  def test_visual_group_add_relationship
-    login
-    visual_group = visual_groups(:visual_group_one)
-    image = images(:peltigera_image)
-    count = VisualGroupImage.count
-    get(:visual_group_status,
-        params: { id: visual_group.id, imgid: image.id, value: "true" })
-    assert_equal(count + 1, VisualGroupImage.count)
-  end
 end
